EE-725: extract ext_ffi & handlers into separate files
