+++ execution-engine/contract-ffi/src/handlers.rs
@@ -1,0 +1,18 @@
+#[panic_handler]
+#[no_mangle]
+pub fn panic(_info: &::core::panic::PanicInfo) -> ! {
+    unsafe {
+        ::core::intrinsics::abort();
+    }
+}
+
+#[alloc_error_handler]
+#[no_mangle]
+pub extern "C" fn oom(_: ::core::alloc::Layout) -> ! {
+    unsafe {
+        ::core::intrinsics::abort();
+    }
+}
+
+#[lang = "eh_personality"]
+extern "C" fn eh_personality() {}
